Use TxHub as cache for all transactions and track tx status

+++ AElf.ChainController/TxMemPool/ITxPoolService.cs
@@ -94,6 +94,6 @@ namespace AElf.ChainController.TxMemPool
         /// <returns></returns>
         Task Revert(List<Transaction> txsOut);
 
-        void SetBlockVolume(ulong minimal, ulong maximal);
+        void SetBlockVolume(int minimal, int maximal);
     }
 }
